:recycle:  move Invocable.java to core
