+++ src/main/java/win/doyto/query/cache/CacheWrapper.java
@@ -1,6 +1,7 @@
 package win.doyto.query.cache;
 
 import org.springframework.cache.Cache;
+import win.doyto.query.core.Invocable;
 
 /**
  * CacheWrapper
