+++ src/main/java/win/doyto/query/core/Invocable.java
@@ -1,4 +1,4 @@
-package win.doyto.query.cache;
+package win.doyto.query.core;
 
 /**
  * Invocable
