+++ src/main/java/win/doyto/query/service/AbstractService.java
@@ -16,11 +16,7 @@ import org.springframework.transaction.support.TransactionCallback;
 import org.springframework.transaction.support.TransactionOperations;
 import org.springframework.transaction.support.TransactionTemplate;
 import win.doyto.query.cache.CacheWrapper;
-import win.doyto.query.cache.Invocable;
-import win.doyto.query.core.DataAccess;
-import win.doyto.query.core.JdbcDataAccess;
-import win.doyto.query.core.MemoryDataAccess;
-import win.doyto.query.core.PageQuery;
+import win.doyto.query.core.*;
 import win.doyto.query.entity.EntityAspect;
 import win.doyto.query.entity.Persistable;
 import win.doyto.query.entity.UserIdProvider;
